+++ plugin-animation/src/main/java/com/mapbox/maps/plugin/animation/HighLevelAnimatorSet.kt
@@ -1,6 +1,7 @@
 package com.mapbox.maps.plugin.animation
 
 import android.animation.AnimatorSet
+import com.mapbox.common.Cancelable
 import com.mapbox.maps.MapboxExperimental
 import com.mapbox.maps.threading.AnimationThreadController