+++ plugin-viewport/src/main/kotlin/com/mapbox/maps/plugin/viewport/state/OverviewViewportStateImpl.kt
@@ -4,9 +4,9 @@ import android.animation.AnimatorSet
 import android.animation.ValueAnimator
 import androidx.annotation.VisibleForTesting
 import androidx.core.animation.doOnEnd
+import com.mapbox.common.Cancelable
 import com.mapbox.maps.CameraOptions
 import com.mapbox.maps.MapboxExperimental
-import com.mapbox.maps.plugin.animation.Cancelable
 import com.mapbox.maps.plugin.animation.camera
 import com.mapbox.maps.plugin.delegates.MapDelegateProvider
 import com.mapbox.maps.plugin.viewport.data.OverviewViewportStateOptions