+++ plugin-viewport/src/test/kotlin/com/mapbox/maps/plugin/viewport/transition/ImmediateViewportTransitionImplTest.kt
@@ -1,9 +1,9 @@
 package com.mapbox.maps.plugin.viewport.transition
 
+import com.mapbox.common.Cancelable
 import com.mapbox.maps.CameraOptions
 import com.mapbox.maps.ScreenCoordinate
 import com.mapbox.maps.plugin.animation.CameraAnimationsPlugin
-import com.mapbox.maps.plugin.animation.Cancelable
 import com.mapbox.maps.plugin.animation.MapAnimationOptions
 import com.mapbox.maps.plugin.animation.camera
 import com.mapbox.maps.plugin.delegates.MapDelegateProvider