+++ sdk-base/src/main/java/com/mapbox/maps/plugin/viewport/state/ViewportState.kt
@@ -1,6 +1,6 @@
 package com.mapbox.maps.plugin.viewport.state
 
-import com.mapbox.maps.plugin.animation.Cancelable
+import com.mapbox.common.Cancelable
 import com.mapbox.maps.plugin.viewport.ViewportPlugin
 import com.mapbox.maps.plugin.viewport.transition.ViewportTransition